+++ src/core/util/String.hpp
@@ -8,7 +8,7 @@ namespace String {
     std::string Join(std::vector<std::string> list, const std::string& delimiter);
     void ReplaceAll(std::string& str, const std::string& from, const std::string& to);
 
-    std::string FileSizeFormat(uint size);
+    std::string FileSizeFormat(size_t size);
     std::string DurationFormat(const sf::Time& time);
 
     bool IsDigit(char ch);
